+++ packages/nocodb/src/db/formulav2/formula-query-builder.helpers.ts
@@ -1,0 +1,33 @@
+import type CustomKnex from '~/db/CustomKnex';
+
+export const getAggregateFn: (
+  fnName: string,
+) => (args: { qb; knex?: CustomKnex; cn }) => any = (parentFn) => {
+  switch (parentFn?.toUpperCase()) {
+    case 'MIN':
+      return ({ qb, cn }) => qb.clear('select').min(cn);
+    case 'MAX':
+      return ({ qb, cn }) => qb.clear('select').max(cn);
+    case 'ADD':
+    case 'SUM':
+    case 'FLOAT':
+    case 'NUMBER':
+    case 'ARITH':
+      return ({ qb, cn }) => qb.clear('select').sum(cn);
+
+    case 'AVG':
+      return ({ qb, cn }) => qb.clear('select').sum(cn);
+
+    // todo:
+    //   return ({ qb, cn, knex, argsCount }) =>
+    //     qb
+    //       .clear('select')
+    //       .select(
+    //         knex.raw('sum(??)/(count(??)) + ?)', [cn, cn, (argsCount || 1) - 1])
+    //       );
+    case 'CONCAT':
+    default:
+      return ({ qb, cn }) => qb.clear('select').concat(cn);
+    // return '';
+  }
+};
