+++ packages/nocodb/src/db/formulav2/formula-query-builder.types.ts
@@ -9,11 +9,11 @@ import type {
   ParsedFormulaNode,
   UnaryExpressionNode,
 } from 'nocodb-sdk';
-import type { BaseModelSqlv2 } from '~/db/BaseModelSqlv2';
 import type { BaseUser, Column, Model, User } from '~/models';
+import type { IBaseModelSqlV2 } from '~/db/IBaseModelSqlV2';
 
 export interface FormulaBaseParams {
-  baseModelSqlv2: BaseModelSqlv2;
+  baseModelSqlv2: IBaseModelSqlV2;
   tableAlias?: string;
   baseUsers?: (Partial<User> & BaseUser)[];
 }
