+++ packages/nocodb/src/db/formulav2/formulaQueryBuilderv2.ts
@@ -1,3 +1,4 @@
+import { Logger } from '@nestjs/common';
 import jsep from 'jsep';
 import {
   ComparisonOperators,
@@ -6,14 +7,13 @@ import {
   JSEPNode,
   LongTextAiMetaProp,
   NcErrorType,
-  RelationTypes,
   UITypes,
   validateDateWithUnknownFormat,
   validateFormulaAndExtractTreeWithType,
 } from 'nocodb-sdk';
-import { Logger } from '@nestjs/common';
-import mapFunctionName from '../mapFunctionName';
 import genRollupSelectv2 from '../genRollupSelectv2';
+import mapFunctionName from '../mapFunctionName';
+import { lookupOrLtarBuilder } from './lookup-or-ltar-builder';
 import type {
   CallExpressionNode,
   ComparisonOperator,
@@ -21,60 +21,25 @@ import type {
   LiteralNode,
   ParsedFormulaNode,
 } from 'nocodb-sdk';
+import type { IBaseModelSqlV2 } from '~/db/IBaseModelSqlV2';
+import type { User } from '~/models';
+import type Column from '~/models/Column';
+import type RollupColumn from '~/models/RollupColumn';
 import type {
   FnParsedTreeNode,
   FormulaQueryBuilderBaseParams,
   TAliasToClumn,
 } from './formula-query-builder.types';
-import type RollupColumn from '~/models/RollupColumn';
-import type LinkToAnotherRecordColumn from '~/models/LinkToAnotherRecordColumn';
-import type LookupColumn from '~/models/LookupColumn';
-import type Column from '~/models/Column';
-import type { User } from '~/models';
-import type { BaseModelSqlv2 } from '~/db/BaseModelSqlv2';
-import type CustomKnex from '~/db/CustomKnex';
-import { BaseUser, ButtonColumn } from '~/models';
-import Model from '~/models/Model';
 import NocoCache from '~/cache/NocoCache';
-import { CacheScope } from '~/utils/globals';
-import { convertDateFormatForConcat } from '~/helpers/formulaFnHelper';
-import FormulaColumn from '~/models/FormulaColumn';
 import { getRefColumnIfAlias } from '~/helpers';
 import { ExternalTimeout, NcBaseErrorv2, NcError } from '~/helpers/catchError';
+import { convertDateFormatForConcat } from '~/helpers/formulaFnHelper';
+import { BaseUser, ButtonColumn } from '~/models';
+import FormulaColumn from '~/models/FormulaColumn';
+import Model from '~/models/Model';
+import { CacheScope } from '~/utils/globals';
 
 const logger = new Logger('FormulaQueryBuilderv2');
-
-export const getAggregateFn: (
-  fnName: string,
-) => (args: { qb; knex?: CustomKnex; cn }) => any = (parentFn) => {
-  switch (parentFn?.toUpperCase()) {
-    case 'MIN':
-      return ({ qb, cn }) => qb.clear('select').min(cn);
-    case 'MAX':
-      return ({ qb, cn }) => qb.clear('select').max(cn);
-    case 'ADD':
-    case 'SUM':
-    case 'FLOAT':
-    case 'NUMBER':
-    case 'ARITH':
-      return ({ qb, cn }) => qb.clear('select').sum(cn);
-
-    case 'AVG':
-      return ({ qb, cn }) => qb.clear('select').sum(cn);
-
-    // todo:
-    //   return ({ qb, cn, knex, argsCount }) =>
-    //     qb
-    //       .clear('select')
-    //       .select(
-    //         knex.raw('sum(??)/(count(??)) + ?)', [cn, cn, (argsCount || 1) - 1])
-    //       );
-    case 'CONCAT':
-    default:
-      return ({ qb, cn }) => qb.clear('select').concat(cn);
-    // return '';
-  }
-};
 
 async function _formulaQueryBuilder(params: FormulaQueryBuilderBaseParams) {
   const {
@@ -188,452 +153,16 @@ async function _formulaQueryBuilder(params: FormulaQueryBuilderBaseParams) {
         break;
       case UITypes.Lookup:
       case UITypes.LinkToAnotherRecord:
-        aliasToColumn[col.id] = async (
-          parentColumns?: Set<string>,
-        ): Promise<any> => {
-          let aliasCount = 0;
-          let selectQb;
-          let isArray = false;
-          const alias = `__nc_formula${aliasCount++}`;
-          const lookup =
-            col.uidt === UITypes.Lookup
-              ? await col.getColOptions<LookupColumn>(context)
-              : null;
-          {
-            const relationCol = lookup
-              ? await lookup.getRelationColumn(context)
-              : col;
-            const relation =
-              await relationCol.getColOptions<LinkToAnotherRecordColumn>(
-                context,
-              );
-            // if (relation.type !== RelationTypes.BELONGS_TO) continue;
-
-            const childColumn = await relation.getChildColumn(context);
-            const parentColumn = await relation.getParentColumn(context);
-            const childModel = await childColumn.getModel(context);
-            await childModel.getColumns(context);
-            const parentModel = await parentColumn.getModel(context);
-            await parentModel.getColumns(context);
-
-            let relationType = relation.type;
-
-            if (relationType === RelationTypes.ONE_TO_ONE) {
-              relationType = relationCol.meta?.bt
-                ? RelationTypes.BELONGS_TO
-                : RelationTypes.HAS_MANY;
-            }
-            let lookupColumn = lookup
-              ? await lookup.getLookupColumn(context)
-              : null;
-
-            switch (relationType) {
-              case RelationTypes.BELONGS_TO:
-                selectQb = knex(
-                  knex.raw(`?? as ??`, [
-                    baseModelSqlv2.getTnPath(parentModel.table_name),
-                    alias,
-                  ]),
-                ).where(
-                  `${alias}.${parentColumn.column_name}`,
-                  knex.raw(`??`, [
-                    `${
-                      tableAlias ??
-                      baseModelSqlv2.getTnPath(childModel.table_name)
-                    }.${childColumn.column_name}`,
-                  ]),
-                );
-                lookupColumn = lookupColumn ?? parentModel.displayValue;
-                break;
-              case RelationTypes.HAS_MANY:
-                isArray = relation.type !== RelationTypes.ONE_TO_ONE;
-                selectQb = knex(
-                  knex.raw(`?? as ??`, [
-                    baseModelSqlv2.getTnPath(childModel.table_name),
-                    alias,
-                  ]),
-                ).where(
-                  `${alias}.${childColumn.column_name}`,
-                  knex.raw(`??`, [
-                    `${
-                      tableAlias ??
-                      baseModelSqlv2.getTnPath(parentModel.table_name)
-                    }.${parentColumn.column_name}`,
-                  ]),
-                );
-                lookupColumn = lookupColumn ?? childModel.displayValue;
-                break;
-              case RelationTypes.MANY_TO_MANY:
-                {
-                  isArray = true;
-                  const mmModel = await relation.getMMModel(context);
-                  const mmParentColumn = await relation.getMMParentColumn(
-                    context,
-                  );
-                  const mmChildColumn = await relation.getMMChildColumn(
-                    context,
-                  );
-
-                  const assocAlias = `__nc${aliasCount++}`;
-                  selectQb = knex(
-                    knex.raw(`?? as ??`, [
-                      baseModelSqlv2.getTnPath(parentModel.table_name),
-                      alias,
-                    ]),
-                  )
-                    .join(
-                      knex.raw(`?? as ??`, [
-                        baseModelSqlv2.getTnPath(mmModel.table_name),
-                        assocAlias,
-                      ]),
-                      `${assocAlias}.${mmParentColumn.column_name}`,
-                      `${alias}.${parentColumn.column_name}`,
-                    )
-                    .where(
-                      `${assocAlias}.${mmChildColumn.column_name}`,
-                      knex.raw(`??`, [
-                        `${
-                          tableAlias ??
-                          baseModelSqlv2.getTnPath(childModel.table_name)
-                        }.${childColumn.column_name}`,
-                      ]),
-                    );
-                  lookupColumn = lookupColumn ?? parentModel.displayValue;
-                }
-                break;
-            }
-
-            let prevAlias = alias;
-            while (lookupColumn.uidt === UITypes.Lookup) {
-              const nestedAlias = `__nc_formula${aliasCount++}`;
-              const nestedLookup =
-                await lookupColumn.getColOptions<LookupColumn>(context);
-              const relationCol = await nestedLookup.getRelationColumn(context);
-              const relation =
-                await relationCol.getColOptions<LinkToAnotherRecordColumn>(
-                  context,
-                );
-              // if any of the relation in nested lookup is
-              // not belongs to then ignore the sort option
-              // if (relation.type !== RelationTypes.BELONGS_TO) continue;
-
-              const childColumn = await relation.getChildColumn(context);
-              const parentColumn = await relation.getParentColumn(context);
-              const childModel = await childColumn.getModel(context);
-              await childModel.getColumns(context);
-              const parentModel = await parentColumn.getModel(context);
-              await parentModel.getColumns(context);
-
-              let relationType = relation.type;
-
-              if (relationType === RelationTypes.ONE_TO_ONE) {
-                relationType = relationCol.meta?.bt
-                  ? RelationTypes.BELONGS_TO
-                  : RelationTypes.HAS_MANY;
-              }
-
-              switch (relationType) {
-                case RelationTypes.BELONGS_TO:
-                  {
-                    selectQb.join(
-                      knex.raw(`?? as ??`, [
-                        baseModelSqlv2.getTnPath(parentModel.table_name),
-                        nestedAlias,
-                      ]),
-                      `${prevAlias}.${childColumn.column_name}`,
-                      `${nestedAlias}.${parentColumn.column_name}`,
-                    );
-                  }
-                  break;
-                case RelationTypes.HAS_MANY:
-                  {
-                    isArray = relation.type !== RelationTypes.ONE_TO_ONE;
-                    selectQb.join(
-                      knex.raw(`?? as ??`, [
-                        baseModelSqlv2.getTnPath(childModel.table_name),
-                        nestedAlias,
-                      ]),
-                      `${prevAlias}.${parentColumn.column_name}`,
-                      `${nestedAlias}.${childColumn.column_name}`,
-                    );
-                  }
-                  break;
-                case RelationTypes.MANY_TO_MANY: {
-                  isArray = true;
-                  const mmModel = await relation.getMMModel(context);
-                  const mmParentColumn = await relation.getMMParentColumn(
-                    context,
-                  );
-                  const mmChildColumn = await relation.getMMChildColumn(
-                    context,
-                  );
-
-                  const assocAlias = `__nc${aliasCount++}`;
-
-                  selectQb
-                    .join(
-                      knex.raw(`?? as ??`, [
-                        baseModelSqlv2.getTnPath(mmModel.table_name),
-                        assocAlias,
-                      ]),
-                      `${assocAlias}.${mmChildColumn.column_name}`,
-                      `${prevAlias}.${childColumn.column_name}`,
-                    )
-                    .join(
-                      knex.raw(`?? as ??`, [
-                        baseModelSqlv2.getTnPath(parentModel.table_name),
-                        nestedAlias,
-                      ]),
-                      `${nestedAlias}.${parentColumn.column_name}`,
-                      `${assocAlias}.${mmParentColumn.column_name}`,
-                    );
-                }
-              }
-
-              /*selectQb.join(
-`${parentModel.title} as ${nestedAlias}`,
-`${nestedAlias}.${parentColumn.title}`,
-`${prevAlias}.${childColumn.title}`
-);*/
-
-              lookupColumn = await nestedLookup.getLookupColumn(context);
-              prevAlias = nestedAlias;
-            }
-
-            switch (lookupColumn.uidt) {
-              case UITypes.Links:
-              case UITypes.Rollup:
-                {
-                  const builder = (
-                    await genRollupSelectv2({
-                      baseModelSqlv2,
-                      knex,
-                      alias: prevAlias,
-                      columnOptions: (await lookupColumn.getColOptions(
-                        context,
-                      )) as RollupColumn,
-                    })
-                  ).builder;
-                  // selectQb.select(builder);
-
-                  if (isArray) {
-                    const qb = selectQb;
-                    selectQb = (fn) =>
-                      knex
-                        .raw(
-                          getAggregateFn(fn)({
-                            qb,
-                            knex,
-                            cn: knex.raw(builder).wrap('(', ')'),
-                          }),
-                        )
-                        .wrap('(', ')');
-                  } else {
-                    selectQb.select(knex.raw(builder).wrap('(', ')'));
-                  }
-                }
-                break;
-              case UITypes.LinkToAnotherRecord:
-                {
-                  const nestedAlias = `__nc_formula${aliasCount++}`;
-                  const relation =
-                    await lookupColumn.getColOptions<LinkToAnotherRecordColumn>(
-                      context,
-                    );
-                  // if (relation.type !== RelationTypes.BELONGS_TO) continue;
-
-                  const colOptions = (await lookupColumn.getColOptions(
-                    context,
-                  )) as LinkToAnotherRecordColumn;
-                  const childColumn = await colOptions.getChildColumn(context);
-                  const parentColumn = await colOptions.getParentColumn(
-                    context,
-                  );
-                  const childModel = await childColumn.getModel(context);
-                  await childModel.getColumns(context);
-                  const parentModel = await parentColumn.getModel(context);
-                  await parentModel.getColumns(context);
-                  let cn;
-
-                  let relationType = relation.type;
-
-                  if (relationType === RelationTypes.ONE_TO_ONE) {
-                    relationType = relationCol.meta?.bt
-                      ? RelationTypes.BELONGS_TO
-                      : RelationTypes.HAS_MANY;
-                  }
-
-                  switch (relationType) {
-                    case RelationTypes.BELONGS_TO:
-                      {
-                        selectQb.join(
-                          knex.raw(`?? as ??`, [
-                            baseModelSqlv2.getTnPath(parentModel.table_name),
-                            nestedAlias,
-                          ]),
-                          `${alias}.${childColumn.column_name}`,
-                          `${nestedAlias}.${parentColumn.column_name}`,
-                        );
-                        cn = knex.raw('??.??', [
-                          nestedAlias,
-                          parentModel?.displayValue?.column_name,
-                        ]);
-                      }
-                      break;
-                    case RelationTypes.HAS_MANY:
-                      {
-                        isArray = relation.type !== RelationTypes.ONE_TO_ONE;
-                        selectQb.join(
-                          knex.raw(`?? as ??`, [
-                            baseModelSqlv2.getTnPath(childModel.table_name),
-                            nestedAlias,
-                          ]),
-                          `${alias}.${parentColumn.column_name}`,
-                          `${nestedAlias}.${childColumn.column_name}`,
-                        );
-                        cn = knex.raw('??.??', [
-                          nestedAlias,
-                          childModel?.displayValue?.column_name,
-                        ]);
-                      }
-                      break;
-                    case RelationTypes.MANY_TO_MANY:
-                      {
-                        isArray = true;
-                        const mmModel = await relation.getMMModel(context);
-                        const mmParentColumn = await relation.getMMParentColumn(
-                          context,
-                        );
-                        const mmChildColumn = await relation.getMMChildColumn(
-                          context,
-                        );
-
-                        const assocAlias = `__nc${aliasCount++}`;
-
-                        selectQb
-                          .join(
-                            knex.raw(`?? as ??`, [
-                              baseModelSqlv2.getTnPath(mmModel.table_name),
-                              assocAlias,
-                            ]),
-                            `${assocAlias}.${mmChildColumn.column_name}`,
-                            `${alias}.${childColumn.column_name}`,
-                          )
-                          .join(
-                            knex.raw(`?? as ??`, [
-                              baseModelSqlv2.getTnPath(parentModel.table_name),
-                              nestedAlias,
-                            ]),
-                            `${nestedAlias}.${parentColumn.column_name}`,
-                            `${assocAlias}.${mmParentColumn.column_name}`,
-                          );
-                      }
-                      cn = knex.raw('??.??', [
-                        nestedAlias,
-                        parentModel?.displayValue?.column_name,
-                      ]);
-                  }
-
-                  selectQb.join(
-                    knex.raw(`?? as ??`, [
-                      baseModelSqlv2.getTnPath(parentModel.table_name),
-                      nestedAlias,
-                    ]),
-                    `${nestedAlias}.${parentColumn.column_name}`,
-                    `${prevAlias}.${childColumn.column_name}`,
-                  );
-
-                  if (isArray) {
-                    const qb = selectQb;
-                    selectQb = (fn) =>
-                      knex
-                        .raw(
-                          getAggregateFn(fn)({
-                            qb,
-                            knex,
-                            cn: lookupColumn.column_name,
-                          }),
-                        )
-                        .wrap('(', ')');
-                  } else {
-                    selectQb.select(`${prevAlias}.${cn}`);
-                  }
-                }
-                break;
-              case UITypes.Formula:
-                {
-                  const formulaOption =
-                    await lookupColumn.getColOptions<FormulaColumn>(context);
-                  const lookupModel = await lookupColumn.getModel(context);
-                  if (parentColumns?.has(lookupColumn.id)) {
-                    NcError.formulaError('Circular reference detected', {
-                      details: {
-                        columnId: lookupColumn.id,
-                        modelId: model.id,
-                        parentColumnIds: Array.from(parentColumns),
-                      },
-                    });
-                  }
-                  const { builder } = await _formulaQueryBuilder({
-                    baseModelSqlv2,
-                    _tree: formulaOption.formula,
-                    model: lookupModel,
-                    aliasToColumn,
-                    parsedTree: formulaOption.getParsedTree(),
-                    parentColumns: new Set([
-                      lookupColumn.id,
-                      ...(parentColumns ?? []),
-                    ]),
-                    tableAlias: prevAlias,
-                  });
-                  if (isArray) {
-                    const qb = selectQb;
-                    selectQb = (fn) =>
-                      knex
-                        .raw(
-                          getAggregateFn(fn)({
-                            qb,
-                            knex,
-                            cn: knex.raw(builder).wrap('(', ')'),
-                          }),
-                        )
-                        .wrap('(', ')');
-                  } else {
-                    selectQb.select(builder);
-                  }
-                }
-                break;
-              default:
-                {
-                  if (isArray) {
-                    const qb = selectQb;
-                    selectQb = (fn) =>
-                      knex
-                        .raw(
-                          getAggregateFn(fn)({
-                            qb,
-                            knex,
-                            cn: `${prevAlias}.${lookupColumn.column_name}`,
-                          }),
-                        )
-                        .wrap('(', ')');
-                  } else {
-                    selectQb.select(`${prevAlias}.${lookupColumn.column_name}`);
-                  }
-                }
-
-                break;
-            }
-
-            if (selectQb)
-              return {
-                builder:
-                  typeof selectQb === 'function'
-                    ? selectQb
-                    : knex.raw(selectQb as any).wrap('(', ')'),
-              };
-          }
-        };
+        aliasToColumn[col.id] = lookupOrLtarBuilder({
+          baseModelSqlv2,
+          context,
+          model,
+          col,
+          _formulaQueryBuilder,
+          knex,
+          tableAlias,
+          aliasToColumn,
+        });
         break;
       case UITypes.Rollup:
       case UITypes.Links:
@@ -1384,7 +913,7 @@ export default async function formulaQueryBuilderv2({
   parsedTree,
   baseUsers,
 }: {
-  baseModel: BaseModelSqlv2;
+  baseModel: IBaseModelSqlV2;
   tree;
   model: Model;
   column?: Column;
